# [IMP] tools: LRU with less locks