Description of the issue/feature this PR addresses:

The most common operation is getting the value for which we would like to avoid locking so that when we run in multi-threaded mode, there are less concurrency issues. We also implement the operations using as much as possible native structures to avoid overhead.

Performance:
```
lru = LRU(50)
%timeit lru[99] = "ok"
%timeit lru[99]
%timeit for i, v in enumerate("hello world"): lru[i] = v

lru = LRU(50)
%timeit lru[randint(1, 60)] = 1
%timeit lru[randint(1, 160)] = 1
%timeit lru[randint(1, 600)] = 1

"""
NEW:

223 ns ± 2.98 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
65.5 ns ± 0.229 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
2.79 μs ± 95.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

496 ns ± 16.6 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
593 ns ± 2.72 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
668 ns ± 6.67 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

OLD:

3.11 μs ± 32.9 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
2.73 μs ± 54.4 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
34.9 μs ± 1.22 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

3.28 μs ± 26.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
3.63 μs ± 265 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
3.6 μs ± 135 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
"""
```


---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr